add tailwind mdx test
