+++ gatsby-browser.js
@@ -1,0 +1,1 @@
+import './src/styles/global.css';
