+++ package.json
@@ -1,1 +1,40 @@
-{"name":"philificent","version":"1.0.0","private":true,"description":"Philificent","author":"Ekxalt","keywords":["gatsby"],"scripts":{"develop":"gatsby develop","start":"gatsby develop","build":"gatsby build","serve":"gatsby serve","clean":"gatsby clean"},"dependencies":{"@mdx-js/mdx":"^1.6.22","@mdx-js/react":"^1.6.22","gatsby":"^2.32.8","gatsby-image":"^2.11.0","gatsby-plugin-manifest":"^2.12.1","gatsby-plugin-mdx":"^1.10.1","gatsby-plugin-offline":"^3.10.2","gatsby-plugin-react-helmet":"^3.10.0","gatsby-plugin-sharp":"^2.14.3","gatsby-plugin-sitemap":"^2.12.0","gatsby-source-filesystem":"^2.11.1","gatsby-transformer-sharp":"^2.12.0","react":"^16.13.1","react-dom":"^16.13.1","react-helmet":"^6.1.0"}}
+{
+  "name": "philificent",
+  "version": "1.0.0",
+  "private": true,
+  "description": "Philificent",
+  "author": "Ekxalt",
+  "keywords": [
+    "gatsby"
+  ],
+  "scripts": {
+    "develop": "gatsby develop",
+    "start": "gatsby develop",
+    "build": "gatsby build",
+    "serve": "gatsby serve",
+    "clean": "gatsby clean"
+  },
+  "dependencies": {
+    "@mdx-js/mdx": "^1.6.22",
+    "@mdx-js/react": "^1.6.22",
+    "gatsby": "^2.32.8",
+    "gatsby-image": "^2.11.0",
+    "gatsby-plugin-manifest": "^2.12.1",
+    "gatsby-plugin-mdx": "^1.10.1",
+    "gatsby-plugin-offline": "^3.10.2",
+    "gatsby-plugin-react-helmet": "^3.10.0",
+    "gatsby-plugin-sharp": "^2.14.3",
+    "gatsby-plugin-sitemap": "^2.12.0",
+    "gatsby-source-filesystem": "^2.11.1",
+    "gatsby-transformer-sharp": "^2.12.0",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
+    "react-helmet": "^6.1.0"
+  },
+  "devDependencies": {
+    "autoprefixer": "^10.2.4",
+    "gatsby-plugin-postcss": "^3.7.0",
+    "postcss": "^8.2.6",
+    "tailwindcss": "^2.0.3"
+  }
+}
