+++ src/pages/updog.mdx
@@ -1,0 +1,19 @@
+---
+title: Sure smells like updog in here
+path: /updog
+date: 2021-02-27
+---
+
+import Layout from "../components/layout"
+
+<section class="mb-8">test in section?</section>
+
+# Nothing, what's up with you?
+
+I miss when life and things were simpler, not like way back pre-electricity times
+no, I'm talking more about the times we used to have where people had "thick-skin",
+and sticking up for someone didn't shame them in ways I still don't understand.
+
+The office was simple once, I liked that.
+
+export default Layout
