+++ src/styles/global.css
@@ -1,0 +1,6 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+#___gatsby, 
+#gatsby-focus-wrapper {height: 100%;}
